View of Study Module Page
Reworked study module feature to be in line with the data the Tests and Flashcards use for navigational purposes. Added simple Study Module view. The state of the study module button changes based on if a study module exists or not. The feature generates a study module and saves it to the database.