Debugged navigation. Fixed study module view on Study Module page and form view for study module on MindMap page
Fixed view of study module page. The name of the study module was not showing. Debugged navigational elements  on the frontend and backend that was making the MindMapPage send the studytool Id instead of the study module Id. Debugged form view on MindMap page where the form disappeared where a test and flashcard set existed, but not a study module.