Debugged navigation Study module generation
Fixed issue where the newly generated study modules were not navigating to the study module page after being generated. Navigating to existing study modules worked fine. Rewrote a conditional statement for it to recognize study module as a type of study tool. The api response was returning the correct data, but I had to correct a property access that was making it assign undefined to the studyModuleId. Tested the the study module feature and it generates a study module and navigates to the newly generated study module. Tested navigation for existing study modules and they work. Tested flashcard and test features and they still work after all of the edits. Navigation works for all three features for existing and generating new study tools.